fix(KDS-754): fix console error from clicking on SVG to close, break out utils into individual files
