+++ packages/time-picker/src/TimePicker/components/index.ts
@@ -1,3 +1,4 @@
+// TODO Rename the subcomponents to be more explicit that they relate to the TimePicker
 export * from "./Button"
 export * from "./DateSegment"
 export * from "./Menu"
