+++ packages/time-picker/src/TimePicker/types.ts
@@ -5,3 +5,12 @@ export type TimeOption = {
   label: string
   value: Time
 }
+export type StatusType = "default" | "error"
+
+export type ValueType = {
+  /**
+   * Supply hour in 24 hour format
+   */
+  hour: number | undefined
+  minutes: number | undefined
+}
