--- packages/time-picker/src/TimePicker/utils/index.ts
@@ -1,1 +1,0 @@
-export * from "./time"
